requirements.txt
apify==2.3.1
langchain-openai==0.3.6
langgraph==0.2.73
aiohttp>=3.8.0
langchain>=0.1.0
pydantic>=2.0.0
langchain-core>=0.1.0
langchain_community==0.3.19
tavily-python==0.5.1